+++ packages/keyring-sdk/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add new entropy support ([#603](https://github.com/MetaMask/accounts/pull/603))
 - Add `encodeMnemonicWords` ([#606](https://github.com/MetaMask/accounts/pull/606))
 
 ### Changed